refactor: 도메인 순환 의존성 해결 (#57)
주요 변경 사항:

1. academy 도메인의 외부 의존성 제거
   - CourseAdapter를 course 도메인으로 이동 (AcademyCourseQueryAdapter)
   - EnrollmentAdapter를 enrollment 도메인으로 이동 (AcademyEnrollmentQueryAdapter)
   - AuthException을 common.exceptions.application.UnauthorizedException으로 교체

2. user → dashboard 의존성 제거
   - LoginSuccessEvent를 user.application.event로 이동

3. board → notification 의존성 제거
   - @NotifyEvent와 NotificationEventType을 common.notification으로 이동
   - 크로스-커팅 관심사를 공통 패키지로 분리

4. course → assessment 의존성 제거
   - AssessmentPort가 course 도메인 전용 타입 사용하도록 변경
   - AssessmentInfo, GradeAssessmentType 도입
   - AssessmentAdapter를 assessment 도메인으로 이동 (CourseGradeAssessmentAdapter)

남은 작업:
- JPA Entity 관계로 인한 순환 의존성 (enrollment ↔ course, attendance ↔ course 등)
- 이들은 DB 스키마와 연결되어 있어 별도 작업 필요
