+++ springboot/src/main/java/com/mzc/backend/lms/domains/academy/adapter/in/web/StudentAcademicTermController.java
@@ -4,7 +4,7 @@ import com.mzc.backend.lms.common.response.ApiResponse;
 import com.mzc.backend.lms.domains.academy.adapter.in.web.dto.AcademicTermResponseDto;
 import com.mzc.backend.lms.domains.academy.application.port.in.AcademicTermQueryUseCase;
 import com.mzc.backend.lms.domains.academy.application.port.in.StudentAcademicTermQueryUseCase;
-import com.mzc.backend.lms.domains.user.exception.AuthException;
+import com.mzc.backend.lms.common.exceptions.application.UnauthorizedException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
@@ -35,7 +35,7 @@ public class StudentAcademicTermController {
     public ResponseEntity<ApiResponse<List<AcademicTermResponseDto>>> listMyAcademicTerms(
             @AuthenticationPrincipal Long studentId) {
         if (studentId == null) {
-            throw AuthException.unauthorized();
+            throw new UnauthorizedException();
         }
         List<AcademicTermResponseDto> data = studentAcademicTermQueryUseCase.getMyAcademicTerms(studentId);
         return ResponseEntity.ok(ApiResponse.success(data));
@@ -48,7 +48,7 @@ public class StudentAcademicTermController {
     public ResponseEntity<ApiResponse<AcademicTermResponseDto>> getCurrentAcademicTerm(
             @AuthenticationPrincipal Long studentId) {
         if (studentId == null) {
-            throw AuthException.unauthorized();
+            throw new UnauthorizedException();
         }
         AcademicTermResponseDto data = academicTermQueryUseCase.getCurrentAcademicTerm(LocalDate.now());
         return ResponseEntity.ok(ApiResponse.success(data));
