+++ springboot/src/main/java/com/mzc/backend/lms/domains/board/application/service/CommentService.java
@@ -14,8 +14,8 @@ import com.mzc.backend.lms.domains.board.exception.BoardException;
 import com.mzc.backend.lms.domains.board.application.port.out.AttachmentRepositoryPort;
 import com.mzc.backend.lms.domains.board.application.port.out.CommentRepositoryPort;
 import com.mzc.backend.lms.domains.board.application.port.out.PostRepositoryPort;
-import com.mzc.backend.lms.domains.notification.aop.annotation.NotifyEvent;
-import com.mzc.backend.lms.domains.notification.aop.event.NotificationEventType;
+import com.mzc.backend.lms.common.notification.annotation.NotifyEvent;
+import com.mzc.backend.lms.common.notification.event.NotificationEventType;
 import com.mzc.backend.lms.domains.user.adapter.in.web.dto.profile.UserBasicInfoDto;
 import com.mzc.backend.lms.domains.user.application.port.in.GetUserInfoCacheUseCase;
 import lombok.RequiredArgsConstructor;
