+++ springboot/src/main/java/com/mzc/backend/lms/domains/course/grade/domain/enums/GradeAssessmentType.java
@@ -1,0 +1,11 @@
+package com.mzc.backend.lms.domains.course.grade.domain.enums;
+
+/**
+ * 성적 계산에 사용하는 평가 유형
+ * assessment 도메인의 AssessmentType과 매핑됨
+ */
+public enum GradeAssessmentType {
+    QUIZ,
+    MIDTERM,
+    FINAL
+}
