+++ springboot/src/main/java/com/mzc/backend/lms/domains/course/notice/adapter/in/event/CourseNoticeNotificationListener.java
@@ -2,7 +2,7 @@ package com.mzc.backend.lms.domains.course.notice.adapter.in.event;
 
 import com.mzc.backend.lms.domains.course.notice.domain.event.CourseNoticeCreatedEvent;
 import com.mzc.backend.lms.domains.enrollment.adapter.out.persistence.repository.EnrollmentRepository;
-import com.mzc.backend.lms.domains.notification.aop.event.NotificationEventType;
+import com.mzc.backend.lms.common.notification.event.NotificationEventType;
 import com.mzc.backend.lms.domains.notification.application.port.in.SendNotificationUseCase;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
